피드백 반영

+++ src/main/java/roomescape/time/ParticipationTimeService.java
@@ -5,6 +5,7 @@ import roomescape.reservation.Reservation;
 import roomescape.reservation.ReservationRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ParticipationTimeService {
@@ -38,10 +39,21 @@ public class ParticipationTimeService {
     }
 
     public ParticipationTime save(ParticipationTime participationTime) {
+
+        Optional<ParticipationTime> existingTime = participationTimeRepository.findByTime(participationTime.getTime());
+
+        if (existingTime.isPresent()) {
+            throw new IllegalArgumentException("이미 존재하는 참여 시간입니다.");
+        }
+
         return participationTimeRepository.save(participationTime);
     }
 
     public void deleteById(Long id) {
+        if (reservationRepository.existsByParticipationTimeId(id)) {
+            throw new IllegalArgumentException("해당 시간에 예약이 존재하여 삭제할 수 없습니다.");
+        }
         participationTimeRepository.deleteById(id);
+
     }
 }
